BasicHeaders - More consistent Reply-To
The CiviMail UI presents two inter-related options:

 * "Track Replies" (override_verp==0)
 * Custom "Reply-To" (override_verp==1; replyto_email defined)

Reading the old logic, it would secretly enable "Track Replies" if the
custom "Reply-To" coincided with the "From".  Weird. Feels like an
unintended consequence of trying to optimize-away duplicate "Reply-To".

Now, it should only "Track Replies" if "Track Replies" is enabled. Sensible.

And it will still optimize-away duplicate "Reply-To".